final code

+++ src/controllers/blogController.js
@@ -1,7 +1,6 @@
 const blogModel = require("../models/blogModel");
 const authorModel = require("../models/authorModel");
 const validation = require("../validator/validation");
-
 let {isEmpty,isValidName, isValidObjectId} = validation;
 
 //___create blogs__________________________________________________________________________________
@@ -10,6 +9,9 @@ let {isEmpty,isValidName, isValidObjectId} = validation;
   try {
     let data = req.body;
     let { title, body, authorId, category } = data;
+    if(Object.keys(data).length == 0){
+      return res.status(400).send({ status: false, msg: "Please provide key in request body" });    
+    }
     if(!isEmpty(authorId)){
       return res.status(400).send({ status: false, msg: "Please provide author ID" });
     } 
@@ -50,7 +52,7 @@ const getBlogs = async function (req, res) {
     const queryParam = req.query;
 
     const { authorId, category, tags, subcategory } = queryParam;
-    if (Object.keys(queryParam) == 0) {
+    if (Object.keys(queryParam).length == 0) {
       return res.status(400).send({ status: false, msg: "InValid request" });
     }
     if (isEmpty(category)) {
@@ -62,7 +64,7 @@ const getBlogs = async function (req, res) {
     }
     if (isEmpty(subcategory)) {
       const subcatArray = subcategory.trim().split(",").map((subcat) => subcat.trim());
-      filterQuery["subcateg"] = { $all: subcatArray };
+      filterQuery["subcategory"] = { $all: subcatArray };
     }
     if (isEmpty(authorId)) {
       filterQuery["authorId"] = authorId;
@@ -82,36 +84,42 @@ const getBlogs = async function (req, res) {
 const updatedBlog = async function (req, res) {
   try {
     const blogId = req.params.blogId;
-    // const authorIdFromToken = req.authorId;
     const blogData = req.body;
 
     let { title, body, tags, subcategory } = blogData;
-    if (!blogId)
+    if (!isValidObjectId(blogId))
       return res.status(404).send({status: false,msg: "Blog Is Not Found , Please Enter Valid Blog Id"});
 
-    if (Object.keys(blogData).length == 0)
+    if (Object.keys(blogData).length == 0){
       return res.status(400).send({ status: false, msg: "Body is required" });
+    }
 
-      // const checkBlog = await blogModel.findOne({blogId:blogId})
-      // if(checkBlog.authorId!==authorIdFromToken){
-      //   res.status(401).send({status:false,msg:"unauthorized access"})
-      // }
+      if(title || title==""){
+        if (!isEmpty(title)) {
+          return res.status(400).send({ status: false, msg: "Please provide title" });
+        }
+        if(!isValidName(title)){
+          return res.status(400).send({ status: false, msg: "title should be alphabets only" });
+        }
+      }
+      if(body || body ==""){
+        if (!isEmpty(body)) {
+          return res.status(400).send({ status: false, msg: "Please provide body of blog" });
+        }
+      }
+      if(tags || tags== ""){
+        if (!isEmpty(category)) {
+          return res.status(400).send({ status: false, msg: "Please provide category" });
+        }
+      }
+      if(subcategory || subcategory== ""){
+        if (!isEmpty(category)) {
+          return res.status(400).send({ status: false, msg: "Please provide category" });
+        }
+      }
 
-    let blog = await blogModel.findOneAndUpdate(
-      { _id: blogId, isDeleted: false },
-      {
-        $set: {
-          isPublished: true,
-          body: body,
-          title: title,
-          publishedAt: new Date(),
-        },
-        $push: { tags: tags, subcategory: subcategory },
-      },
-      { new: true }
-    );
-
-    return res.status(200).send({ status: true, data: blog });
+    let blog = await blogModel.findOneAndUpdate({ _id: blogId, isDeleted: false },{$set: {isPublished: true,body: body,title: title,publishedAt: new Date()},$push: { tags: tags, subcategory: subcategory }},{ new: true });
+    return res.status(200).send({ status: true, msg: "Blog updated successfully",data: blog });
   } catch (error) {
     return res.status(500).send({ status: false, Error: error.message });
   }
@@ -124,62 +132,32 @@ const deletedBlog = async (req, res) => {
     let checkBlogId = await blogModel.findById(blogId);
 
     if(!checkBlogId || (checkBlogId.isDeleted == true)){
-      return res.status(404).send({
-          status : false,
-          msg : "Blog has been deleted successfully"
-      })         
+      return res.status(404).send({status : false,msg : "Blog has been deleted successfully"})         
   }
-     let deletedBlog = await blogModel.findOneAndUpdate(
-      { _id: blogId },
-      { $set: { isDeleted: true, deletedAt: Date.now() } },
-      { new: true }
-    );
-    return res.status(200).send({
-      status: true,
-      msg: "Blog has been deleted successfully",
-      data: deletedBlog,
-    });
-  } catch (err) {
-    return res.status(500).send({ status: false, msg: err.message });
-  }
+     let deletedBlog = await blogModel.findOneAndUpdate({ _id: blogId },{ $set: { isDeleted: true,deletedAt: Date.now()}},{ new: true });
+    return res.status(200).send({status: true,msg: "Blog has been deleted successfully",data:deletedBlog});
+  } catch (err) {return res.status(500).send({ status: false, msg: err.message })}
 };
 
 //______delete blogs api 2 by given fields _________________________________________>>>
 
 const deleteByQueryParams = async function (req, res) {
-
   try {
-
-    let data = req.query
-      const deleteByQuery = await blogModel.updateMany(
-
-      { $and: [data ,{authorId : req.id}, { isDeleted: false }] },
-
-      { $set: { isDeleted: true ,deletedAt:new Date()} },
-
-      { new: true, upsert : true })
-
-      if (deleteByQuery.modifiedCount==0) 
-      return res.status(404).send(
-        { status: false,
-           msg: "No Blog Found"
-         })
+    let data = req.query;
+    if(!Object.keys(data).length==0){
+      return res.status(400).send({status: false, message: 'keys must be present'});
+    }
+      const deleteByQuery = await blogModel.updateMany({ $and: [data ,{authorId : req.id}, { isDeleted: false }] },{ $set: { isDeleted: true ,deletedAt:new Date()} },{ new: true, upsert : true })
       let count = deleteByQuery.modifiedCount
-      res.status(200).send({
-          status : true,
-          msg : "No of blogs deleted:", count
-      })
+      if (deleteByQuery.modifiedCount==0) {
+      return res.status(404).send({ status: false, msg: "No Blog Found"})
   }
+      res.status(200).send({status : true, msg : "No of blogs deleted:", count })
+  }
+  catch (err) {res.status(500).send({status:false,msg: err.message})}
+};
 
-  catch (err) {
-      console.log("Error is from delete blogs by query", err.message)
-      res.status(500).send({
-        status:false,
-        msg: err.message 
-      })}
-}
-
-//==================================module exporting =========================================================
+//======================module exporting ==================================
 
 module.exports.createBlog = createBlog;
 module.exports.getBlogs = getBlogs;
